+++ docs/HISTORY.rst
@@ -5,8 +5,12 @@ Changelog
 0.2 (unreleased)
 ~~~~~~~~~~~~~~~~
 
+ - Makes sure bogus urls don't lead to an exception. Fixes #3
+   [href]
+
  - Fixes KML before loading, to avoid common problems found in external
    KML sources.
+   [href]
 
 0.1 (2014-10-29)
 ~~~~~~~~~~~~~~~~